Automating marking unstable tests stable
This commit would automate the process of marking unstable tests stable if the test passes for 4 days
and has a failure probability of zero. This would be the reverse process of flag failing. Tests that
would fail in the flag failing pipeline would be used to check if a test has been marked unstable. Based
on that, the test would then be added to the "unstable data-base". The tests in the database would then
be run and failure probabilities would be calculated. The first occurance would also be in the data base
using which it would be checked if any of the tests have a failure probability of zero and have been so
for four days in which case the test would be marked stable. The commit part would be catered over in the
builder part which would be done in a subsequent commit.

Patch-4: Addressing AliR's comments

Change-Id: I70507c4954a882c18635922d7f6296cb5bf9a23c
Signed-off-by: Irfan Ali Shah <irfans@plumgrid.com>