Cronjob for daily cost usage of all users and all pipelines.
The cronjob would send out an email with the daily and month usage report listing
the cost and the time usage of each user and each pipeline.
Following is a template of the email that would be recieved:

Hi All,
The cost for aurora instances for the past day/month is as follows:

Cost for all pipelines are as follows:

Cost for pipeline1 = $4.180 Approximate Duration 19:25:0
Cost for pipeline2 = $.880 Approximate Duration 4:52:0
Cost for pipeline3 = $0 Approximate Duration 0:0:0
Cost for pipeline4 = $.220 Approximate Duration 1:1:0

Cost for all users are as follows:

Cost for user1 = $5.280 Approximate Duration 24:0:0
Cost for user2 = $7.920 Approximate Duration 24:0:0
Cost for user3 = $.720 Approximate Duration 0:0:0
Cost for user4 = $10.560 Approximate Duration 48:0:0

The other cronjob scripts would be deleted in the subsequent commit.

Change-Id: I297526d1ac2bdc2c8eb2cdb51e623d022e3e23be
Signed-off-by: Irfan Ali Shah <irfans@plumgrid.com>